requirements.txt
diagrams
streamlit
openai
google-generativeai
graphviz
